# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/backstage/admin/login": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backstage"
                ],
                "summary": "Backstage login",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.LoginDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.LoginResponseDTO"
                        }
                    }
                }
            }
        },
        "/backstage/cache": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Forestage"
                ],
                "summary": "Cache View",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/cache/delete/{cacheName}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Forestage"
                ],
                "summary": "Cache Any Delete",
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/carousel": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Carousel"
                ],
                "summary": "Carousel View",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "int default",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            15,
                            20,
                            30,
                            40,
                            50
                        ],
                        "type": "integer",
                        "description": "int enums",
                        "name": "pageLimit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sort",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "id",
                            "key"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "sortColumn",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "",
                        "description": "string default",
                        "name": "searchCategory",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.CarouselListDTO"
                        }
                    }
                }
            }
        },
        "/backstage/carousel/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Carousel"
                ],
                "summary": "Carousel Create",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.CarouselCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/carousel/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Carousel"
                ],
                "summary": "Carousel Delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/carousel/edit/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Carousel"
                ],
                "summary": "Carousel Edit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.CarouselCreateOrEditDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/backstage/carousel/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Backstage-Carousel"
                ],
                "summary": "Carousel By Id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.CarouselIdDTO"
                        }
                    }
                }
            }
        },
        "/backstage/jwt/check": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backstage"
                ],
                "summary": "Backstage CheckToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "string",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.JwtUserInfoDTO"
                        }
                    }
                }
            }
        },
        "/backstage/jwt/refreshtoken": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Backstage"
                ],
                "summary": "Backstage RefreshToken",
                "parameters": [
                    {
                        "description": "json",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backstagedto.JwtRefTokenDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backstagedto.LoginResponseDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backstagedto.CarouselCreateOrEditDTO": {
            "type": "object",
            "properties": {
                "endTime": {
                    "description": "結束時間",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "圖片名稱",
                    "type": "string"
                },
                "picture": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forestagedto.PictureListData"
                    }
                },
                "startTime": {
                    "description": "開始時間",
                    "type": "string"
                },
                "status": {
                    "description": "狀態(開關)",
                    "type": "boolean"
                },
                "weight": {
                    "description": "權重",
                    "type": "integer"
                }
            }
        },
        "backstagedto.CarouselData": {
            "type": "object",
            "properties": {
                "createTime": {
                    "description": "新增時間",
                    "type": "string"
                },
                "createUserId": {
                    "description": "新增人員",
                    "type": "integer"
                },
                "endTime": {
                    "description": "結束時間",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "圖片名稱",
                    "type": "string"
                },
                "startTime": {
                    "description": "開始時間",
                    "type": "string"
                },
                "status": {
                    "description": "狀態(開關)",
                    "type": "boolean"
                },
                "updateTime": {
                    "description": "更新時間",
                    "type": "string"
                },
                "updateUserId": {
                    "description": "修改人員",
                    "type": "integer"
                },
                "weight": {
                    "description": "權重",
                    "type": "integer"
                }
            }
        },
        "backstagedto.CarouselIdDTO": {
            "type": "object",
            "properties": {
                "carousel": {
                    "$ref": "#/definitions/backstagedto.CarouselData"
                },
                "picture": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/forestagedto.PictureListData"
                    }
                }
            }
        },
        "backstagedto.CarouselListDTO": {
            "type": "object",
            "properties": {
                "carousel": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backstagedto.CarouselData"
                    }
                },
                "pageData": {
                    "$ref": "#/definitions/dto.PageForMultSearchDTO"
                }
            }
        },
        "backstagedto.JwtRefTokenDTO": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "backstagedto.JwtTokenDTO": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "backstagedto.JwtUserInfoDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "backstagedto.LoginDTO": {
            "type": "object",
            "properties": {
                "loginName": {
                    "type": "string",
                    "minLength": 4
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
        "backstagedto.LoginResponseDTO": {
            "type": "object",
            "properties": {
                "authorityJwt": {
                    "$ref": "#/definitions/backstagedto.JwtTokenDTO"
                },
                "userInfo": {
                    "$ref": "#/definitions/backstagedto.JwtUserInfoDTO"
                }
            }
        },
...
        "dto.PageForMultSearchDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageLimit": {
                    "type": "integer"
                },
                "search": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "sort": {
                    "type": "string"
                },
                "sortColumn": {
                    "type": "string"
                }
            }
        },
...
        "forestagedto.PictureListData": {
            "type": "object",
            "properties": {
                "alt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pictureUrl": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "url": {
                    "type": "string"
                },
                "weight": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}